moved components arround about, feel free to mess with this how ever you please
