+++ frontend/src/components/App.jsx
@@ -2,7 +2,7 @@ import { BrowserRouter as Router,Routes, Route } from 'react-router-dom';
 import Login from './Login'
 import DashBoard from './DashBoard'
 import NotFound from './NotFound'
-import { UserProvider } from './userContext'
+import { UserProvider } from '../userContext'
 
 function Home(){
   return(
